[Refactor][OPS] update manifest source.op for elementwise split
Update 66 manifest entries' source.op to point at the new per-cluster
files under tileops/ops/elementwise/. Move LerpTensorFwdOp into
arithmetic.py so it shares source.op with its variant_of primary
LerpFwdOp (R16 cross-entry consistency).
